+++ changes.rst
@@ -14,6 +14,7 @@ Bug fixes:
    expressions.
 -  Fix ``string_length`` function to return the number of Unicode code points
    rather than the number of bytes (i.e., to interpret the string as UTF-8).
+-  Emit type error for opt tuples and records created through type aliases.
 
 Changes:
 ^^^^^^^^